requirements.txt
openai>=1.0.0
anthropic>=0.20.0
python-dotenv>=1.0.0
pytest>=7.0.0
streamlit>=1.28.0